+++ src/ui/TorrentInfoDialog.cpp
@@ -2,6 +2,7 @@
 #include "../core/Item.h"
 #include "../core/TorrentParser.h"
 #include "../core/TrackerClient.h"
+#include <QAbstractItemView>
 #include <QDateTime>
 #include <QFormLayout>
 #include <QHBoxLayout>
@@ -86,7 +87,8 @@ void TorrentInfoDialog::setupUi() {
     QLabel *filesLabel = new QLabel("<b>Files:</b>");
     mainLayout->addWidget(filesLabel);
 
-    QTableWidget *filesTable = new QTableWidget(m_info.files.size(), 2, this);
+    QTableWidget *filesTable =
+        new QTableWidget(static_cast<int>(m_info.files.size()), 2, this);
     filesTable->setHorizontalHeaderLabels({"Path", "Size (MB)"});
     filesTable->horizontalHeader()->setSectionResizeMode(0,
                                                          QHeaderView::Stretch);
@@ -96,7 +98,7 @@ void TorrentInfoDialog::setupUi() {
     filesTable->setEditTriggers(QAbstractItemView::NoEditTriggers);
     filesTable->setMaximumHeight(150);
 
-    for (int i = 0; i < m_info.files.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(m_info.files.size()); ++i) {
       filesTable->setItem(i, 0, new QTableWidgetItem(m_info.files[i].path));
       double fileMb =
           static_cast<double>(m_info.files[i].length) / (1024.0 * 1024.0);
@@ -108,7 +110,8 @@ void TorrentInfoDialog::setupUi() {
     mainLayout->addWidget(filesTable);
   }
 
-  m_trackerTable = new QTableWidget(m_info.trackers.size(), 5, this);
+  m_trackerTable =
+      new QTableWidget(static_cast<int>(m_info.trackers.size()), 5, this);
   m_trackerTable->setHorizontalHeaderLabels(
       {"Tracker", "Status", "Seeders", "Leechers", "Downloaded"});
   m_trackerTable->horizontalHeader()->setSectionResizeMode(
@@ -118,7 +121,7 @@ void TorrentInfoDialog::setupUi() {
   m_trackerTable->setSelectionBehavior(QAbstractItemView::SelectRows);
   m_trackerTable->setEditTriggers(QAbstractItemView::NoEditTriggers);
 
-  for (int i = 0; i < m_info.trackers.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(m_info.trackers.size()); ++i) {
     QTableWidgetItem *trackerItem = new QTableWidgetItem(m_info.trackers[i]);
     trackerItem->setFlags(Qt::ItemIsUserCheckable | Qt::ItemIsEnabled |
                           Qt::ItemIsSelectable);
@@ -147,8 +150,8 @@ void TorrentInfoDialog::setupUi() {
     historyTable->setShowGrid(false);
 
     QJsonArray history = m_item->metadata["history"].toArray();
-    historyTable->setRowCount(history.size());
-    for (int i = 0; i < history.size(); ++i) {
+    historyTable->setRowCount(static_cast<int>(history.size()));
+    for (int i = 0; i < static_cast<int>(history.size()); ++i) {
       QJsonObject entry = history[i].toObject();
       QDateTime dt =
           QDateTime::fromString(entry["timestamp"].toString(), Qt::ISODate);
@@ -251,7 +254,7 @@ void TorrentInfoDialog::processNextTracker() {
     m_currentTrackerIndex++;
   }
 
-  if (m_currentTrackerIndex >= m_info.trackers.size()) {
+  if (m_currentTrackerIndex >= static_cast<int>(m_info.trackers.size())) {
     m_isQuerying = false;
     m_queryBtn->setEnabled(true);
     m_cancelBtn->setEnabled(false);
